clear out old ddm and call ddm-next ddm
