--- ddm-next-admin-client/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "ddm-next-admin-client"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-slog = "2.7"
-anyhow = "1.0"
-percent-encoding = "2.1"
-reqwest = { version = "0.11", features = ["json", "stream", "rustls-tls"] }
-progenitor = { git = "https://github.com/oxidecomputer/progenitor" }
