--- ddm-next-admin-client/src/lib.rs
@@ -1,14 +1,0 @@
-progenitor::generate_api!(
-    spec = "../ddm-openapi/ddm-next-admin.json",
-    inner_type = slog::Logger,
-    pre_hook = (|log: &slog::Logger, request: &reqwest::Request| {
-        slog::debug!(log, "client request";
-            "method" => %request.method(),
-            "uri" => %request.url(),
-            "body" => ?&request.body(),
-        );
-    }),
-    post_hook = (|log: &slog::Logger, result: &Result<_, _>| {
-        slog::debug!(log, "client response"; "result" => ?result);
-    }),
-);
