--- ddm-next/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "ddm-next"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-slog = { version = "2.7.0", features = ["max_level_trace", "release_max_level_trace"] }
-socket2 = { version = "0.4.7", features = ["all"] }
-ispf = { git = "https://github.com/oxidecomputer/ispf" }
-serde = { version = "1.0.150", features = ["derive"] }
-hostname = "0.3.1"
-thiserror = "1.0.37"
-dropshot = { git = "https://github.com/oxidecomputer/dropshot", branch = "main" }
-schemars = { version = "0.8.0", features = [ "uuid" ] }
-tokio = { version = "1.21", features = ["full"] }
-serde_repr = "0.1.9"
-slog-envlogger = "2.2.0"
-anyhow = "1.0.66"
-hyper = "0.14.23"
-serde_json = "1.0.89"
-libnet = { git = "https://github.com/oxidecomputer/netadm-sys", branch = "ddm" }
-
-[dependencies.dpd-api]
-git = "https://github.com/oxidecomputer/dendrite"
-branch = "tfport-softnpu"
-package = "dpd-api"
-
-[dependencies.dendrite-common]
-git = "https://github.com/oxidecomputer/dendrite"
-branch = "tfport-softnpu"
-package = "common"
