--- ddm-next/src/bin/ddm-next-apigen.rs
@@ -1,12 +1,0 @@
-use anyhow::anyhow;
-use anyhow::Result;
-use ddm_next::admin::api_description;
-use std::fs::File;
-
-fn main() -> Result<()> {
-    let api = api_description().map_err(|e| anyhow!("{}", e))?;
-    let openapi = api.openapi("DDM Admin", "v0.1.0");
-    let mut out = File::create("ddm-admin.json")?;
-    openapi.write(&mut out)?;
-    Ok(())
-}
