--- ddm-openapi/ddm-next-admin.json
@@ -1,235 +1,0 @@
-{
-  "openapi": "3.0.3",
-  "info": {
-    "title": "DDM Admin",
-    "version": "v0.1.0"
-  },
-  "paths": {
-    "/peers": {
-      "get": {
-        "operationId": "get_peers",
-        "responses": {
-          "200": {
-            "description": "successful operation",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "title": "Map_of_PeerInfo",
-                  "type": "object",
-                  "additionalProperties": {
-                    "$ref": "#/components/schemas/PeerInfo"
-                  }
-                }
-              }
-            }
-          },
-          "4XX": {
-            "$ref": "#/components/responses/Error"
-          },
-          "5XX": {
-            "$ref": "#/components/responses/Error"
-          }
-        }
-      }
-    },
-    "/prefix": {
-      "put": {
-        "operationId": "advertise_prefixes",
-        "requestBody": {
-          "content": {
-            "application/json": {
-              "schema": {
-                "title": "Set_of_Ipv6Prefix",
-                "type": "array",
-                "items": {
-                  "$ref": "#/components/schemas/Ipv6Prefix"
-                },
-                "uniqueItems": true
-              }
-            }
-          },
-          "required": true
-        },
-        "responses": {
-          "204": {
-            "description": "resource updated"
-          },
-          "4XX": {
-            "$ref": "#/components/responses/Error"
-          },
-          "5XX": {
-            "$ref": "#/components/responses/Error"
-          }
-        }
-      },
-      "delete": {
-        "operationId": "withdraw_prefixes",
-        "requestBody": {
-          "content": {
-            "application/json": {
-              "schema": {
-                "title": "Set_of_Ipv6Prefix",
-                "type": "array",
-                "items": {
-                  "$ref": "#/components/schemas/Ipv6Prefix"
-                },
-                "uniqueItems": true
-              }
-            }
-          },
-          "required": true
-        },
-        "responses": {
-          "204": {
-            "description": "resource updated"
-          },
-          "4XX": {
-            "$ref": "#/components/responses/Error"
-          },
-          "5XX": {
-            "$ref": "#/components/responses/Error"
-          }
-        }
-      }
-    },
-    "/prefixes": {
-      "get": {
-        "operationId": "get_prefixes",
-        "responses": {
-          "200": {
-            "description": "successful operation",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "title": "Map_of_Set_of_Ipv6Prefix",
-                  "type": "object",
-                  "additionalProperties": {
-                    "type": "array",
-                    "items": {
-                      "$ref": "#/components/schemas/Ipv6Prefix"
-                    },
-                    "uniqueItems": true
-                  }
-                }
-              }
-            }
-          },
-          "4XX": {
-            "$ref": "#/components/responses/Error"
-          },
-          "5XX": {
-            "$ref": "#/components/responses/Error"
-          }
-        }
-      }
-    },
-    "/sync": {
-      "put": {
-        "operationId": "sync",
-        "responses": {
-          "204": {
-            "description": "resource updated"
-          },
-          "4XX": {
-            "$ref": "#/components/responses/Error"
-          },
-          "5XX": {
-            "$ref": "#/components/responses/Error"
-          }
-        }
-      }
-    }
-  },
-  "components": {
-    "responses": {
-      "Error": {
-        "description": "Error",
-        "content": {
-          "application/json": {
-            "schema": {
-              "$ref": "#/components/schemas/Error"
-            }
-          }
-        }
-      }
-    },
-    "schemas": {
-      "Error": {
-        "description": "Error information from a response.",
-        "type": "object",
-        "properties": {
-          "error_code": {
-            "type": "string"
-          },
-          "message": {
-            "type": "string"
-          },
-          "request_id": {
-            "type": "string"
-          }
-        },
-        "required": [
-          "message",
-          "request_id"
-        ]
-      },
-      "Ipv6Prefix": {
-        "type": "object",
-        "properties": {
-          "addr": {
-            "type": "string",
-            "format": "ipv6"
-          },
-          "len": {
-            "type": "integer",
-            "format": "uint8",
-            "minimum": 0
-          }
-        },
-        "required": [
-          "addr",
-          "len"
-        ]
-      },
-      "PeerInfo": {
-        "type": "object",
-        "properties": {
-          "addr": {
-            "type": "string",
-            "format": "ipv6"
-          },
-          "host": {
-            "type": "string"
-          },
-          "kind": {
-            "$ref": "#/components/schemas/RouterKind"
-          },
-          "status": {
-            "$ref": "#/components/schemas/PeerStatus"
-          }
-        },
-        "required": [
-          "addr",
-          "host",
-          "kind",
-          "status"
-        ]
-      },
-      "PeerStatus": {
-        "type": "string",
-        "enum": [
-          "NoContact",
-          "Active",
-          "Expired"
-        ]
-      },
-      "RouterKind": {
-        "type": "integer",
-        "enum": [
-          0,
-          1
-        ]
-      }
-    }
-  }
-}
