+++ ddm/Cargo.toml
@@ -4,23 +4,21 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-icmpv6 = { path = "../icmpv6" }
-libnet = { git = "https://github.com/oxidecomputer/netadm-sys", branch = "ddm" }
-tokio = { version = "1.21", features = ["full"] }
-socket2 = { version = "0.4", features = ["all"] }
-anyhow = "1"
-thiserror = "1"
-util = { path = "../util" }
-slog = "2.7"
-slog-term = "2.7"
-slog-async = "2.7"
-slog-envlogger = "2.2"
+slog = { version = "2.7.0", features = ["max_level_trace", "release_max_level_trace"] }
+socket2 = { version = "0.4.7", features = ["all"] }
+ispf = { git = "https://github.com/oxidecomputer/ispf" }
+serde = { version = "1.0.150", features = ["derive"] }
+hostname = "0.3.1"
+thiserror = "1.0.37"
 dropshot = { git = "https://github.com/oxidecomputer/dropshot", branch = "main" }
 schemars = { version = "0.8.0", features = [ "uuid" ] }
-serde = "1.0"
-serde_json = "1.0"
-hyper = { version = "0.14.20", features = ["full"] }
-futures = "0.3.24"
+tokio = { version = "1.21", features = ["full"] }
+serde_repr = "0.1.9"
+slog-envlogger = "2.2.0"
+anyhow = "1.0.66"
+hyper = "0.14.23"
+serde_json = "1.0.89"
+libnet = { git = "https://github.com/oxidecomputer/netadm-sys", branch = "ddm" }
 
 [dependencies.dpd-api]
 git = "https://github.com/oxidecomputer/dendrite"
