--- ddm/src/net.rs
@@ -1,45 +1,0 @@
-use std::net::AddrParseError;
-use std::net::Ipv6Addr;
-use std::num::ParseIntError;
-use std::str::FromStr;
-
-use schemars::JsonSchema;
-use serde::Deserialize;
-use serde::Serialize;
-use thiserror::Error;
-
-#[derive(
-    Debug, Clone, Copy, Deserialize, Serialize, JsonSchema, PartialEq, Eq, Hash,
-)]
-pub struct Ipv6Prefix {
-    pub addr: Ipv6Addr,
-    pub mask: u8,
-}
-
-#[derive(Debug, Error)]
-pub enum Ipv6PrefixParseError {
-    #[error("expected CIDR representation <addr>/<mask>")]
-    Cidr,
-
-    #[error("address parse error: {0}")]
-    Addr(#[from] AddrParseError),
-
-    #[error("mask parse error: {0}")]
-    Mask(#[from] ParseIntError),
-}
-
-impl FromStr for Ipv6Prefix {
-    type Err = Ipv6PrefixParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let parts: Vec<&str> = s.split('/').collect();
-        if parts.len() < 2 {
-            return Err(Ipv6PrefixParseError::Cidr);
-        }
-
-        Ok(Ipv6Prefix {
-            addr: Ipv6Addr::from_str(parts[0])?,
-            mask: u8::from_str(parts[1])?,
-        })
-    }
-}
