+++ ddm/src/util.rs
