--- ddmadm-next/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "ddmadm-next"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { version = "4.0.29", features = ["derive"] }
-ddm-next = { path = "../ddm-next" }
-ddm-next-admin-client = { path = "../ddm-next-admin-client" }
-slog = "2.7.0"
-slog-term = "2.9.0"
-slog-envlogger = "2.2.0"
-slog-async = "2.7.0"
-tokio = "1.23.0"
-tabwriter = { version = "1", features = ["ansi_formatting"] }
-colored = "2.0.0"
-anyhow = "1.0.67"
