+++ ddmd/Cargo.toml
@@ -4,17 +4,17 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-once_cell = "1.10"
-tokio = { version = "1.21", features = ["full"] }
-slog = "2.7"
-slog-term = "2.7"
-slog-async = "2.7"
-slog-envlogger = "2.2"
-ddm = { path = "../ddm" }
-hostname = "0.3"
 clap = { version = "4.0.29", features = ["derive"] }
+ddm = { path = "../ddm" }
+libnet = { git = "https://github.com/oxidecomputer/netadm-sys", branch = "ddm" }
+slog = "2.7.0"
+slog-async = "2.7.0"
+slog-envlogger = "2.2.0"
+slog-term = "2.9.0"
+tokio = "1.23.0"
+ctrlc = { version = "3.2.4", features = ["termination"] }
 
 [dependencies.dpd-api]
 git = "https://github.com/oxidecomputer/dendrite"
 package = "dpd-api"
-branch = "tfport-softnpu"
+branch = "main"
