--- icmpv6/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "icmpv6"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = "1.0"
-schemars = { version = "0.8.0", features = [ "uuid" ] }
